AI_DETECTION/requirements.txt
cv2
tensorflow
numpy
matplotlib.pyplot
os